+++ saleor/plugins/avatax/tests/test_avatax.py
@@ -1705,6 +1705,7 @@ def test_order_created(api_post_request_task_mock, order, plugin_configuration):
         from_city="WROCŁAW",
         from_postal_code="53-601",
         from_country="PL",
+        shipping_tax_code="FR00001",
     )
     conf = {data["name"]: data["value"] for data in plugin_conf.configuration}
 
@@ -1758,6 +1759,7 @@ def test_order_created(api_post_request_task_mock, order, plugin_configuration):
         "from_postal_code": conf["from_postal_code"],
         "from_country": conf["from_country"],
         "from_country_area": conf["from_country_area"],
+        "shipping_tax_code": conf["shipping_tax_code"],
     }
 
     api_post_request_task_mock.assert_called_once_with(
@@ -2101,11 +2103,19 @@ def test_get_checkout_lines_data_sets_different_tax_code_for_zero_amount(
         checkout_with_item, lines, [], get_plugins_manager()
     )
 
-    # when
-    lines_data = get_checkout_lines_data(
-        checkout_info,
-        lines,
+    config = AvataxConfiguration(
+        username_or_account="test",
+        password_or_license="test",
+        use_sandbox=False,
+        from_street_address="Tęczowa 7",
+        from_city="WROCŁAW",
+        from_country_area="",
+        from_postal_code="53-601",
+        from_country="PL",
     )
+
+    # when
+    lines_data = get_checkout_lines_data(checkout_info, lines, config)
 
     # then
     assert lines_data[0]["amount"] == "0.00"
@@ -2135,11 +2145,19 @@ def test_get_checkout_lines_data_sets_different_tax_code_only_for_zero_amount(
         checkout_with_item, lines, [], get_plugins_manager()
     )
 
-    # when
-    lines_data = get_checkout_lines_data(
-        checkout_info,
-        lines,
+    config = AvataxConfiguration(
+        username_or_account="test",
+        password_or_license="test",
+        use_sandbox=False,
+        from_street_address="Tęczowa 7",
+        from_city="WROCŁAW",
+        from_country_area="",
+        from_postal_code="53-601",
+        from_country="PL",
     )
+
+    # when
+    lines_data = get_checkout_lines_data(checkout_info, lines, config)
 
     # then
     assert lines_data[0]["amount"] == "11.00"
@@ -2162,8 +2180,19 @@ def test_get_order_lines_data_sets_different_tax_code_for_zero_amount(
     )
     variant.product.save()
 
+    config = AvataxConfiguration(
+        username_or_account="test",
+        password_or_license="test",
+        use_sandbox=False,
+        from_street_address="Tęczowa 7",
+        from_city="WROCŁAW",
+        from_country_area="",
+        from_postal_code="53-601",
+        from_country="PL",
+    )
+
     # when
-    lines_data = get_order_lines_data(order_with_lines)
+    lines_data = get_order_lines_data(order_with_lines, config)
 
     # then
     assert lines_data[0]["amount"] == "0.000"
@@ -2187,8 +2216,19 @@ def test_get_order_lines_data_sets_different_tax_code_only_for_zero_amount(
     )
     variant.product.save()
 
+    config = AvataxConfiguration(
+        username_or_account="test",
+        password_or_license="test",
+        use_sandbox=False,
+        from_street_address="Tęczowa 7",
+        from_city="WROCŁAW",
+        from_country_area="",
+        from_postal_code="53-601",
+        from_country="PL",
+    )
+
     # when
-    lines_data = get_order_lines_data(order_with_lines)
+    lines_data = get_order_lines_data(order_with_lines, config)
 
     # then
     assert lines_data[0]["amount"] == "10.000"